MASSIVE PROJECT REWRITE!!!
Completely rewrote D-Wire. This project no longer requires Factorio Server Manager to host a Factorio server anymore. Its completely stand alone and will install and manage a Factorio server for you.

Rewrote the sendchat py and called it discord_to_server, because it essentially sends all chat from the discord server to the game.

downloadsaves rewritten to allow a user to use a drop down menu to pick a save file , download it or remove it.

mods.py rewritten to work without FSM api, while still logging into factorio mods portal to download mods, it automatically installs mods to the server and writes the appropriate settings to the mods list file.

mpsettings.py is new and allows you to set the server presence settings via discord.

settings.py is new and allows you to set your server settings to allow your server to show up in the public server browser list.

readlog.py now uses a json config,

save.py now uploads a file directly to the server and checks for overwriting issues

server_management.py has had a complete overhaul, it handles its own STDout verbose logging, instance locking, discord bot status which matches server status and allows the server to continue to run even if the discord bot dies.

status.py no longer calls the fsm api and gets the server status directly from the verbose log

some changes in the config.json to support the changes in the cogs

Server_info.txt is where the server command line and pid is stored, do not delete this file.

We dont use a country file for the mmdb, upgraded to city, which is capable of storing country as well.